refactor: :zap: leiva's petitions
