--- src/pages/speakers.tsx
@@ -1,35 +1,0 @@
-/* eslint-disable @next/next/no-img-element */
-import type { NextPage } from 'next'
-import Head from 'next/head'
-import GridPonentes from '../components/gridPonentes'
-import { Moleculas } from '../components/moleculas'
-import { Navbar } from '../components/navbar'
-import styles from '../styles/speakersPage.module.css'
-
-const Home: NextPage = () => {
-  
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Salmorejo Tech 2023</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/images/favicon.ico" />
-        <meta charSet="UTF-8" />
-        <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1"/>
-        
-        {/** Meta Info General*/}
-        <meta name="description" content="SalmorejoTech es un evento tecnológico para todos aquellos dispuestos a aprender" />
-        <meta name="keywords"
-          content="evento, ASL, salmorejo, salmorejotech, tech, salmorejotech2023, 2023, salmorejotech2022, 2022, salmorejotech2021, congreso, mesa redonda, conferencia, libre, software libre" />
-        <link rel="canonical" href="https://salmorejo.tech" />
-      </Head>
-
-      <main>
-        <GridPonentes/>
-      </main>
-    </div>
-  )
-}
-
-export default Home
